refactor: split base class and element definition
